Fix errors on WIN32 build
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@97598 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: f3f62769b3ed517c54ad95792c74810233ff560b

+++ Online/Online/PyDIM/src/pydim_utils.cpp
@@ -629,7 +629,7 @@ iterator_to_buffer(PyObject     *iter,   /* list or tuple PyObject */
 	  goto shortbuffer;
 	tmp1 = PyNumber_Float(tmp);
 	if (!tmp1) goto invalid_format;
-	f = PyFloat_AsDouble(tmp1);
+	f = (float)PyFloat_AsDouble(tmp1);
 	memcpy(&buffer[buf_ptr], &f, _DIM_FLOAT_LEN);
 	buf_ptr += _DIM_FLOAT_LEN;                         
 	break;
